+++ app/src/main/java/woowacourse/omok/repository/OmokRepository.kt
@@ -3,8 +3,8 @@ package woowacourse.omok.repository
 import android.content.ContentValues
 import android.database.Cursor
 import android.database.sqlite.SQLiteDatabase
-import woowacourse.omok.database.OmokConstact.TABLE_COLUMN_ID
-import woowacourse.omok.database.OmokConstact.TABLE_NAME
+import woowacourse.omok.database.OmokContract.TABLE_COLUMN_ID
+import woowacourse.omok.database.OmokContract.TABLE_NAME
 
 class OmokRepository(val database: SQLiteDatabase) {
